Rewrite to iterative with a stack instead of recursive
I found another shortcoming of Matej's otherwise-amazing deep
ungroup extension: when you patch inkex to use the XML_PARSE_HUGE
option, you can hit the Python function recursion depth limit in
the "propagate_attribs" function.  I've restructured it to be
iterative with a stack instead of recursive.  The "_ungroup"
function is also recursive, and needs similar treatment.

There was also a nasty performance bug in the interaction between
the propagate_attribs and _ungroup functions, causing
propagate_attribs to be called applied multiple times per node:
Because the _ungroup function propagates the attribs from the
current object to the children, and the propagate_attribs
function recursively propagates the attribs all the way down to
the leaf nodes before returning, the objects at the bottom of the
nested grouping have the propagate_attribs function called on
them as many times as there is depth of nested grouping.  If you
have millions of elements that are thousands of levels of nesting
deep, this is a HUGE slowdown.

I've incorporated the deep_utils.py and the deep_ungroup.py to do
it all in one sweep through the hierarchy, touching each node
only once, with no recursion, meaning it now runs in time O(N)
and space O(log N) with a low coefficient.  Before it was running
in time O(N log N) and space O (log N) with a high coefficient,
so this means a huge speedup with a small space decrease.

It is now trivial to add a "max depth of X" limit (add current
depth as a member of the tuple in the queue and only ungroup if
it's less than X), and possible to have a "leave the last
Y levels" limit (perform the ungroup Y levels up from the current
level instead of at the current level), though I've not
implemented those features.